Fix nokogiri rpath build.
This is a revisit of the earlier Gentoo fix, but is relevant for any
system that might be strict on linking paths and doesn't have the
nokogiri lib in the ld.so search path.

Firstly, the initial linking will fail. Simply fixing that is NOT
sufficent on it's own (you can see with ldd .../nokogumbo.so), as it
will not update the RPATH in the ELF output, and that needs to also be
specified.

We conditionally wrap this using have_library, so it's only attempted if
actually needed.

This might still cause breakages if they have a statically linked
nokogiri.

Just -L + -l leads to:
/usr/lib64/ruby/site_ruby/2.1.0/rubygems/core_ext/kernel_require.rb:55:in `require': nokogiri.so: cannot open shared object file: No such file or directory - .../nokogumbo/ext/nokogumboc/nokogumboc.so (LoadError)

$ ldd ./ext/nokogumboc/nokogumboc.so
  linux-vdso.so.1 (0x00007ffd2db70000)
  nokogiri.so => not found
  libruby21.so.2.1 => /usr/lib64/libruby21.so.2.1 (0x00007f565c4e7000)
  libxml2.so.2 => /usr/lib64/libxml2.so.2 (0x00007f565c17d000)
  ...

With -L + -l + -Wl,-rpath, it works, and ldd output:
$ ldd ext/nokogumboc/nokogumboc.so
  linux-vdso.so.1 (0x00007ffd55f7b000)
  nokogiri.so => /usr/lib64/ruby/gems/2.1.0/gems/nokogiri-1.6.8.1/ext/nokogiri/nokogiri.so (0x00007f823ab55000)
  libruby21.so.2.1 => /usr/lib64/libruby21.so.2.1 (0x00007f823a656000)
  libxml2.so.2 => /usr/lib64/libxml2.so.2 (0x00007f823a2ec000)

Fixes: https://github.com/rubys/nokogumbo/issues/40
Fixes: https://github.com/rubys/nokogumbo/pull/47
Signed-off-by: Robin H. Johnson <robbat2@gentoo.org>